requirements.txt
python-telegram-bot
psycopg2-binary
python-dotenv
google-api-python-client
beautifulsoup4
aiohttp
redis